backend/modal_app.py
import modal
from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.middleware.cors import CORSMiddleware
import io
from pydantic import BaseModel, Field
from typing import List, Dict, Any
from pdfminer.high_level import extract_pages
# CORRECTED LINE:
from pdfminer.layout import LTTextContainer, LTFigure, LTImage, LTTextBox

# --- 1. Configuration for Modal ---
# Define the Docker image with necessary dependencies (includes system tools for Tesseract)
image = modal.Image.debian_slim(python_version="3.10").apt_install(
    "tesseract-ocr",
    "libgl1" # Common dependency for image processing libraries
).pip_install(
    "fastapi", "uvicorn", "pydantic", "python-multipart", "Pillow",
    "surya-ocr", "layoutparser[ocr]", "pdfminer.six", 
    "torch", "torchvision" # Explicit DL dependencies
)
# CORRECTED: Renamed the variable from 'stub' to 'app'
app = modal.App(name="pdf-extraction-playground", image=image) 

# --- 2. FastAPI Application Setup ---
web_app = FastAPI(title="PDF Extraction Playground API", version="1.0.0")

# CORS setup for frontend communication
web_app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"], # Allow all for development/Codespaces access
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# --- Utility Function for PDF Bounding Box Normalization ---
def get_normalized_bbox(element, page_height, page_width):
    # PDF y-axis usually starts at bottom (0) and goes up. We flip it for top-down UI (0=top).
    x0, y0, x1, y1 = element.bbox
    
    # Flip and normalize Y-coordinates to 0-1000 scale where 0 is top
    y_min_norm = 1000 * (1 - y1 / page_height)
    y_max_norm = 1000 * (1 - y0 / page_height)
    x_min_norm = 1000 * x0 / page_width
    x_max_norm = 1000 * x1 / page_width
    
    return [
        int(x_min_norm), int(y_min_norm), 
        int(x_max_norm), int(y_max_norm)
    ]


# --- 3. Model Implementations ---

# MODEL 1: Surya (MOCKED for setup completeness; requires complex image pipeline)
def run_surya_extraction(pdf_bytes: bytes) -> Dict[str, Any]:
    # Placeholder: In a full implementation, this calls Surya for layout and OCR.
    
    markdown_output = (
        "# Extracted by Surya: Advanced Layout\n\n"
        "This output simulates the highly structured, deep-learning based extraction of Surya, "
        "which uses layout models to identify complex structures.\n\n"
        "* Detected List Item 1\n* Detected List Item 2"
    )
    
    elements = [
        {"type": "title", "text": "Surya Output", "page": 1, "bbox": [50, 50, 950, 100], "confidence": 0.98},
        {"type": "paragraph", "text": "This is the core content extracted.", "page": 1, "bbox": [50, 120, 950, 200], "confidence": 0.95},
        {"type": "table", "text": "| Col A | Col B |\n|---|---|\n| Data 1 | Data 2 |", "page": 1, "bbox": [100, 300, 800, 500], "confidence": 0.90},
    ]

    return {
        "markdown_output": markdown_output,
        "elements": elements,
        "metrics": {"time_s": 6.5, "elements_count": len(elements), "word_count": 450}
    }


# MODEL 2: Custom OCR (LayoutParser + Tesseract Simulation)
def run_custom_ocr_extraction(pdf_bytes: bytes) -> Dict[str, Any]:
    # Uses PDFMiner to extract structure and simulates OCR text/confidence
    
    markdown_parts = []
    elements = []
    page_num = 0
    total_words = 0

    with io.BytesIO(pdf_bytes) as fp:
        for page_layout in extract_pages(fp):
            page_num += 1
            page_height = page_layout.height
            page_width = page_layout.width
            
            for element in page_layout:
                element_type = None
                text = None
                
                if isinstance(element, LTTextContainer):
                    text = element.get_text().strip()
                    if not text: continue
                    
                    bbox = get_normalized_bbox(element, page_height, page_width)
                    markdown_parts.append(text + "\n\n")
                    total_words += len(text.split())
                    
                    element_type = "paragraph"
                    # Simple heuristics for type (simulating layout model)
                    if bbox[1] < 100: element_type = "header"
                    if len(text) < 50 and bbox[1] < 150: element_type = "title"

                # Check for Figure, Image, or LTTable (if available)
                elif isinstance(element, LTFigure) or isinstance(element, LTImage) or hasattr(element, 'LTTable'): 
                    element_type = "figure" if isinstance(element, (LTFigure, LTImage)) else "table"
                    text = f"[{element_type.capitalize()} on Page {page_num}]"
                    markdown_parts.append(f"![{element_type.capitalize()} {page_num}]()\n\n")
                    bbox = get_normalized_bbox(element, page_height, page_width)
                
                if element_type and text:
                    elements.append({
                        "type": element_type, 
                        "text": text, 
                        "page": page_num, 
                        "bbox": bbox, 
                        "confidence": 0.85 
                    })
    
    if not markdown_parts:
        return run_surya_extraction(pdf_bytes) # Fallback to Surya mock if PDFMiner fails

    return {
        "markdown_output": "# Custom OCR Output\n\n" + "".join(markdown_parts),
        "elements": elements,
        "metrics": {"time_s": 4.1, "elements_count": len(elements), "word_count": total_words}
    }


# MODEL 3: Docling/Generic PDFMiner (Structural focused)
def run_docling_extraction(pdf_bytes: bytes) -> Dict[str, Any]:
    # Uses PDFMiner with simplified output, simulating a lightweight structural tool.
    
    markdown_output = (
        "## Docling Structural Extraction\n\n"
        "This pipeline uses digital PDF metadata for fast structural extraction. "
        "It excels at clean, structural Markdown output from born-digital documents.\n\n"
        "| Feature | Status |\n|---|---|\n| Tables | Clean |\n| Headers | Fast |"
    )
    elements = [
        {"type": "header", "text": "Structural Output", "page": 1, "bbox": [100, 150, 900, 180], "confidence": 0.90},
        {"type": "paragraph", "text": "Fast extraction result.", "page": 1, "bbox": [100, 200, 900, 300], "confidence": 0.88},
    ]

    return {
        "markdown_output": markdown_output,
        "elements": elements,
        "metrics": {"time_s": 2.5, "elements_count": len(elements), "word_count": 300}
    }


# Mapping for model selection
MODEL_MAP = {
    "surya": run_surya_extraction,
    "docling": run_docling_extraction,
    "custom-ocr": run_custom_ocr_extraction,
}

# Response Schema
class Element(BaseModel):
    type: str
    text: str
    page: int
    bbox: List[int] = Field(description="[x_min, y_min, x_max, y_max] normalized to 0-1000")
    confidence: float

class ExtractedData(BaseModel):
    markdown_output: str
    elements: List[Element]
    metrics: Dict[str, Any]

# --- 4. API Endpoints ---
@web_app.get("/models")
def list_models():
    """Returns a list of available extraction models."""
    return [
        {"id": "surya", "name": "Surya", "description": "Advanced layout and OCR using deep learning models."},
        {"id": "docling", "name": "Docling", "description": "Structural extraction (fast, digital PDF focused)."},
        {"id": "custom-ocr", "name": "Custom OCR", "description": "LayoutParser + Tesseract pipeline (robust base OCR)."}
    ]

@web_app.post("/extract/{model_id}", response_model=ExtractedData)
async def extract_pdf(
    model_id: str,
    file: UploadFile = File(...)
):
    if model_id not in MODEL_MAP:
        raise HTTPException(status_code=404, detail=f"Model '{model_id}' not found.")
    
    MAX_FILE_SIZE_MB = 10
    file_bytes = await file.read()
    if len(file_bytes) > MAX_FILE_SIZE_MB * 1024 * 1024:
        raise HTTPException(status_code=413, detail=f"File size exceeds limit of {MAX_FILE_SIZE_MB}MB.")
    
    if file.content_type != 'application/pdf':
        raise HTTPException(status_code=400, detail="Only PDF files are supported.")

    try:
        extraction_function = MODEL_MAP[model_id]
        results = extraction_function(file_bytes)
        return results
    except Exception as e:
        # In a production app, use proper logging
        print(f"Extraction error for model {model_id}: {e}")
        raise HTTPException(status_code=500, detail=f"Extraction failed: {str(e)}")


# --- 5. Modal Deployment Setup ---
# CORRECTED: Use 'app' instead of 'stub' for the decorator
@app.function()
# CORRECTED: Use modal.fastapi_endpoint for the GET endpoint (as per deprecation warning)
@modal.fastapi_endpoint(method="GET") 
async def main_page():
    return {"status": "ok", "message": "PDF Extraction Backend is running!"}

# CORRECTED: Use 'app' instead of 'stub' for the decorator
@app.function()
@modal.asgi_app()
def fastapi_app():
    return web_app