backend/requirements.txt
fastapi
uvicorn
modal
python-multipart
Pillow
surya-ocr
layoutparser[ocr]
pdfminer.six
torch
torchvision
